fix: bug fixes and prevention
* Circular dependencies for share functions resolved.
* Toast border added for separation between directly stacked.
* Toast context adjusted, method for limiting the selected toasts added.
* Bug fix isAreas computation.
