+++ src/routes/dealers/DealerItem.tsx
@@ -1,19 +1,17 @@
-import { captureException } from "@sentry/react-native";
 import { useTranslation } from "react-i18next";
-import { Share, StyleSheet } from "react-native";
+import { StyleSheet } from "react-native";
 import { ScrollView } from "react-native-gesture-handler";
 
+import { shareDealer } from "./Dealers.common";
 import { appStyles } from "../../components/AppStyles";
 import { DealerContent } from "../../components/dealers/DealerContent";
 import { Floater, padFloater } from "../../components/generic/containers/Floater";
 import { Header } from "../../components/generic/containers/Header";
-import { appBase, conAbbr } from "../../configuration";
 import { useAppRoute } from "../../hooks/nav/useAppNavigation";
 import { useUpdateSinceNote } from "../../hooks/records/useUpdateSinceNote";
 import { useLatchTrue } from "../../hooks/util/useLatchTrue";
 import { useAppSelector } from "../../store";
 import { dealersSelectors } from "../../store/eurofurence/selectors/records";
-import { DealerDetails } from "../../store/eurofurence/types";
 
 /**
  * Params handled by the screen in route.
@@ -43,13 +41,3 @@ export const DealerItem = () => {
         </ScrollView>
     );
 };
-
-export const shareDealer = (dealer: DealerDetails) =>
-    Share.share(
-        {
-            title: dealer.FullName,
-            url: `${appBase}/Web/Dealers/${dealer.Id}`,
-            message: `Check out ${dealer.FullName} on ${conAbbr}!\n${appBase}/Web/Dealers/${dealer.Id}`,
-        },
-        {},
-    ).catch(captureException);
